client/requirements.txt
# Websockets.io
python-socketio[client]